+++ core/screenshot-testing/src/main/kotlin/com/android/geto/core/screenshottesting/util/ScreenshotHelper.kt
@@ -18,15 +18,25 @@
 package com.android.geto.core.screenshottesting.util
 
 import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.key
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.test.junit4.AndroidComposeTestRule
 import androidx.compose.ui.test.onRoot
 import androidx.test.ext.junit.rules.ActivityScenarioRule
+import com.android.geto.core.designsystem.theme.GetoTheme
 import com.github.takahirom.roborazzi.ExperimentalRoborazziApi
 import com.github.takahirom.roborazzi.RoborazziOptions
 import com.github.takahirom.roborazzi.RoborazziOptions.CompareOptions
 import com.github.takahirom.roborazzi.RoborazziOptions.RecordOptions
 import com.github.takahirom.roborazzi.captureRoboImage
+import com.google.accompanist.testharness.TestHarness
+import org.robolectric.RuntimeEnvironment
 
 @OptIn(ExperimentalRoborazziApi::class)
 val DefaultRoborazziOptions = RoborazziOptions(
@@ -49,21 +59,27 @@ fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.c
     fileName: String,
     deviceName: String,
     deviceSpec: String,
+    roborazziOptions: RoborazziOptions = DefaultRoborazziOptions,
     darkMode: Boolean = false,
     body: @Composable () -> Unit,
 ) {
-    captureScreenForDevice(
-        fileName = fileName,
-        deviceName = deviceName,
-        deviceSpec = deviceSpec,
-        darkMode = darkMode,
-        body = body,
-        onCapture = { filePath, roborazziOptions ->
-            onRoot().captureRoboImage(
-                filePath = filePath,
-                roborazziOptions = roborazziOptions,
-            )
-        },
+    val (width, height, dpi) = extractSpecs(deviceSpec)
+
+    RuntimeEnvironment.setQualifiers("w${width}dp-h${height}dp-${dpi}dpi")
+
+    activity.setContent {
+        CompositionLocalProvider(
+            LocalInspectionMode provides true,
+        ) {
+            TestHarness(darkMode = darkMode) {
+                body()
+            }
+        }
+    }
+
+    onRoot().captureRoboImage(
+        filePath = "src/test/screenshots/${fileName}_$deviceName.png",
+        roborazziOptions = roborazziOptions,
     )
 }
 
@@ -88,19 +104,48 @@ fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.c
 fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.captureScreenMultiTheme(
     name: String,
     overrideFileName: String? = null,
+    roborazziOptions: RoborazziOptions = DefaultRoborazziOptions,
     content: @Composable (description: String) -> Unit,
 ) {
-    captureScreenMultiTheme(
-        name = name,
-        overrideFileName = overrideFileName,
-        content = content,
-        onCapture = { filePath, roborazziOptions ->
-            onRoot().captureRoboImage(
-                filePath = filePath,
-                roborazziOptions = roborazziOptions,
-            )
-        },
-    )
+    var description by mutableStateOf("")
+
+    var greenTheme by mutableStateOf(false)
+
+    var purpleTheme by mutableStateOf(false)
+
+    var darkTheme by mutableStateOf(false)
+
+    var dynamicTheme by mutableStateOf(false)
+
+    this.setContent {
+        CompositionLocalProvider(
+            LocalInspectionMode provides true,
+        ) {
+            GetoTheme(
+                greenTheme = greenTheme,
+                purpleTheme = purpleTheme,
+                darkTheme = darkTheme,
+                dynamicTheme = dynamicTheme,
+            ) {
+                key(greenTheme, purpleTheme, darkTheme, dynamicTheme) {
+                    content(description)
+                }
+            }
+        }
+    }
+
+    swrMultiThemes.forEach { (swrGreenTheme, swrPurpleTheme, swrDarkTheme, swrDynamicTheme, swrDescription) ->
+        description = swrDescription
+        greenTheme = swrGreenTheme
+        purpleTheme = swrPurpleTheme
+        darkTheme = swrDarkTheme
+        dynamicTheme = swrDynamicTheme
+
+        onRoot().captureRoboImage(
+            filePath = "src/test/screenshots/" + "$name/${overrideFileName ?: name}" + swrDescription + ".png",
+            roborazziOptions = roborazziOptions,
+        )
+    }
 }
 
 /**
@@ -116,3 +161,56 @@ internal fun extractSpecs(deviceSpec: String): TestDeviceSpecs {
 }
 
 data class TestDeviceSpecs(val width: Int, val height: Int, val dpi: Int)
+
+private data class SwrMultiTheme(
+    val greenTheme: Boolean,
+    val purpleTheme: Boolean,
+    val darkTheme: Boolean,
+    val dynamicTheme: Boolean,
+    val description: String,
+)
+
+private val swrMultiThemes = listOf(
+    SwrMultiTheme(
+        greenTheme = true,
+        purpleTheme = false,
+        darkTheme = false,
+        dynamicTheme = false,
+        description = "GreenTheme",
+    ),
+    SwrMultiTheme(
+        greenTheme = false,
+        purpleTheme = true,
+        darkTheme = false,
+        dynamicTheme = false,
+        description = "PurpleTheme",
+    ),
+    SwrMultiTheme(
+        greenTheme = true,
+        purpleTheme = false,
+        darkTheme = true,
+        dynamicTheme = false,
+        description = "GreenDarkTheme",
+    ),
+    SwrMultiTheme(
+        greenTheme = false,
+        purpleTheme = true,
+        darkTheme = true,
+        dynamicTheme = false,
+        description = "PurpleDarkTheme",
+    ),
+    SwrMultiTheme(
+        greenTheme = false,
+        purpleTheme = false,
+        darkTheme = false,
+        dynamicTheme = true,
+        description = "DynamicTheme",
+    ),
+    SwrMultiTheme(
+        greenTheme = false,
+        purpleTheme = false,
+        darkTheme = true,
+        dynamicTheme = true,
+        description = "DynamicDarkTheme",
+    ),
+)
